--- qadence2_expressions/core/__init__.py
@@ -1,51 +1,0 @@
-from __future__ import annotations
-
-from .ircompiler import irc
-from .primitives import (
-    RX,
-    RY,
-    RZ,
-    Z0,
-    Z1,
-    X,
-    Xm,
-    Xp,
-    Y,
-    Z,
-    NOT,
-    cos,
-    exp,
-    function,
-    set_grid_scale,
-    set_grid_type,
-    set_qubits_positions,
-    sin,
-    symbol,
-    unitary_hermitian_operator,
-    value,
-)
-
-__all__ = [
-    "irc",
-    "value",
-    "symbol",
-    "exp",
-    "function",
-    "unitary_hermitian_operator",
-    "X",
-    "Y",
-    "Z",
-    "Z0",
-    "Z1",
-    "Xp",
-    "Xm",
-    "RX",
-    "RY",
-    "RZ",
-    "NOT",
-    "sin",
-    "cos",
-    "set_qubits_positions",
-    "set_grid_type",
-    "set_grid_scale",
-]
