+++ qadence2_expressions/functions.py
@@ -1,0 +1,27 @@
+from __future__ import annotations
+
+from .expression import Expression
+from .constructors import function, value
+
+
+def sin(x: Expression | complex | float | int) -> Expression:
+    if not isinstance(x, Expression):
+        x = value(x)
+    return function("sin", x)
+
+
+def cos(x: Expression | complex | float | int) -> Expression:
+    if not isinstance(x, Expression):
+        x = value(x)
+    return function("cos", x)
+
+
+# Exponential function as power.
+def exp(x: Expression | complex | float | int) -> Expression:
+    return Expression.symbol("E") ** x
+
+
+def log(x: Expression | complex | float | int) -> Expression:
+    if not isinstance(x, Expression):
+        x = value(x)
+    return function("log", x)
